+++ src/main/java/com/example/callaccountingsystem/repository/PricingUnitRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PricingUnitRepository extends JpaRepository<PricingUnitEntity, Long> {
+public interface PricingUnitRepository extends JpaRepository<PricingUnitEntity, Integer> {
 
     Page<PricingUnitEntity> findAll(Pageable pageable);
 }
