+++ src/main/java/com/example/callaccountingsystem/repository/TariffPlanRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TariffPlanRepository extends JpaRepository<TariffPlanEntity, Long> {
+public interface TariffPlanRepository extends JpaRepository<TariffPlanEntity, Integer> {
 
     Page<TariffPlanEntity> findAll(Pageable pageable);
 }
